+++ storage/tianmu/core/delta_record_head.h
@@ -54,7 +54,7 @@ class DeltaRecordHeadForInsert : public DeltaRecordHead {
   size_t field_count_;
   utils::BitSet null_mask_;
 
-  int64_t *field_head_;
+  uint32_t *field_head_;
   int null_offset_;
   int field_offset_;
 };
@@ -76,7 +76,7 @@ class DeltaRecordHeadForUpdate : public DeltaRecordHead {
   size_t field_count_;
   utils::BitSet update_mask_;
   utils::BitSet null_mask_;
-  int64_t * field_head_;
+  uint32_t *field_head_;
   int null_offset_;
   int update_offset_;
   int field_offset_;
